+++ fluxdataqaqc/__init__.py
@@ -1,10 +1,10 @@
 """
-Tools for performing energy balance correction and data validation of eddy covariance flux station data. Includes routines for adjusting near surface latent energy and sensible heat fluxes to improve closure of the energy balance, calculation of evapotranspiration, and potential clear sky radiation. Other functionalities include data processing e.g. temporal resampling and aggregation, unit conversions, cleaning data using quality control metrics, linear regression routines, and interactive visualizations or input and output.
+Tools for performing energy balance correction and QA/QC of eddy covariance flux station data. For example, routines for adjusting near surface latent energy and sensible heat fluxes to improve closure of the energy balance. Other functionalities include data processing e.g. temporal resampling, interpolation, and aggregation, unit conversions, cleaning data using quality control metrics, linear regression, meterological calcuations, and interactive visualizations.
 """
 
 __name__ = 'fluxdataqaqc'
 __author__ = 'John Volk'
-__version__ = '0.1.1'
+__version__ = '0.1.2'
 
 
 from fluxdataqaqc.data import Data